+++ src/main/java/de/samply/exporter/ExporterServiceException.java
@@ -1,0 +1,13 @@
+package de.samply.exporter;
+
+public class ExporterServiceException extends Exception{
+
+    public ExporterServiceException(String message) {
+        super(message);
+    }
+
+    public ExporterServiceException(Throwable cause) {
+        super(cause);
+    }
+
+}